+++ sdk/servicebus/service-bus/src/core/managementClient.ts
@@ -134,9 +134,10 @@ const correlationProperties = [
 
 /**
  * @internal
+ * @ignore
  * Options to set when updating the disposition status
  */
-interface DispositionStatusOptions extends OperationOptions {
+export interface DispositionStatusOptions extends OperationOptions {
   /**
    * @property [propertiesToModify] A dictionary of Service Bus brokered message properties
    * to modify.
@@ -393,10 +394,12 @@ export class ManagementClient extends LinkEntity {
    */
   async close(): Promise<void> {
     try {
+      // Always clear the timeout, as the isOpen check may report
+      // false without ever having cleared the timeout otherwise.
+      clearTimeout(this._tokenRenewalTimer as NodeJS.Timer);
       if (this._isMgmtRequestResponseLinkOpen()) {
         const mgmtLink = this._mgmtReqResLink;
         this._mgmtReqResLink = undefined;
-        clearTimeout(this._tokenRenewalTimer as NodeJS.Timer);
         await mgmtLink!.close();
         log.mgmt("Successfully closed the management session.");
       }