+++ sdk/servicebus/service-bus/test/renewLock.spec.ts
@@ -30,7 +30,7 @@ describe("renew lock", () => {
     receiverClient = await serviceBusClient.test.getPeekLockReceiver(entityNames);
 
     senderClient = serviceBusClient.test.addToCleanup(
-      serviceBusClient.createSender(entityNames.queue ?? entityNames.topic!)
+      await serviceBusClient.createSender(entityNames.queue ?? entityNames.topic!)
     );
   }